Refactored base rules to improve readability.
